// src/lib/db/types.ts
export type NbUsersRow = {
  id: string;
  name: string;
  hubspot_owner_id: string;
  owner_name: string;
  pin_salt: string;
  pin_hash: string;
  prefs_id: number;
  features: string[] | null;
  created_at: string;
  disabled_at: string | null;
};